chore: rename generateEntropy to entropyGenerator
