+++ lib/util/entropyGenerator.js
